Send subtitle separately from tracks
Allows us to remove PostEditingData.java in favor of a consolidated EditingData.java
