+++ modules/editor-service-api/src/main/java/org/opencastproject/editor/api/EditingData.java
@@ -22,6 +22,7 @@ package org.opencastproject.editor.api;
 
 import static java.util.Objects.requireNonNull;
 
+import org.opencastproject.mediapackage.MediaPackageElementFlavor;
 import org.opencastproject.util.data.Tuple;
 
 import com.google.gson.Gson;
@@ -46,10 +47,11 @@ public class EditingData {
   @SerializedName(WORKFLOW_ACTIVE)
   private final Boolean workflowActive;
   private final List<String> waveformURIs;
+  private final List<Subtitle> subtitles;
 
   public EditingData(List<SegmentData> segments, List<TrackData> tracks, List<WorkflowData> workflows, Long duration,
           String title, String recordingStartDate, String seriesId, String seriesName, Boolean workflowActive,
-          List<String> waveformURIs) {
+          List<String> waveformURIs, List<Subtitle> subtitles) {
     this.segments = segments;
     this.tracks = tracks;
     this.workflows = workflows;
@@ -59,12 +61,13 @@ public class EditingData {
     this.series = new SeriesData(seriesId, seriesName);
     this.workflowActive = workflowActive;
     this.waveformURIs = waveformURIs;
+    this.subtitles = subtitles;
   }
 
-  public static PostEditingData parse(String json) {
+  public static EditingData parse(String json) {
     requireNonNull(json);
     Gson gson = new Gson();
-    PostEditingData editingData = gson.fromJson(json, PostEditingData.class);
+    EditingData editingData = gson.fromJson(json, EditingData.class);
     requireNonNull(editingData.getTracks());
     requireNonNull(editingData.getSegments());
 
@@ -90,9 +93,31 @@ public class EditingData {
     return Collections.unmodifiableList(tracks);
   }
 
+  public List<Subtitle> getSubtitles() {
+    return subtitles;
+  }
+
   public String toString() {
     Gson gson = new GsonBuilder().serializeNulls().create();
     return gson.toJson(this);
+  }
+
+  public static final class Subtitle {
+    private final MediaPackageElementFlavor flavor;
+    private final String subtitle;
+
+    public Subtitle(MediaPackageElementFlavor flavor, String subtitle) {
+      this.flavor = flavor;
+      this.subtitle = subtitle;
+    }
+
+    public MediaPackageElementFlavor getFlavor() {
+      return flavor;
+    }
+
+    public String getSubtitle() {
+      return subtitle;
+    }
   }
 }
